+++ src/apps/qi-dao/polygon/qi-dao.escrowed-qi.contract-position-fetcher.ts
@@ -1,64 +1,33 @@
 import { Inject } from '@nestjs/common';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
-import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getTokenImg } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { ContractPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { GetTokenBalancesParams, GetTokenDefinitionsParams } from '~position/template/contract-position.template.types';
+import { VotingEscrowTemplateContractPositionFetcher } from '~position/template/voting-escrow.template.contract-position-fetcher';
 
-import { QiDaoContractFactory } from '../contracts';
-import { QI_DAO_DEFINITION } from '../qi-dao.definition';
+import { QiDaoContractFactory, QiDaoEscrowedQi } from '../contracts';
 
-const appId = QI_DAO_DEFINITION.id;
-const groupId = QI_DAO_DEFINITION.groups.escrowedQi.id;
-const network = Network.POLYGON_MAINNET;
+@PositionTemplate()
+export class PolygonQiDaoEscrowedQiContractPositionFetcher extends VotingEscrowTemplateContractPositionFetcher<QiDaoEscrowedQi> {
+  groupLabel = 'Escrowed QI';
+  veTokenAddress = '0x880decade22ad9c58a8a4202ef143c4f305100b3';
 
-@Register.ContractPositionFetcher({ appId, groupId, network })
-export class PolygonQiDaoEscrowedQiContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(QiDaoContractFactory) private readonly qiDaoContractFactory: QiDaoContractFactory,
-  ) {}
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(QiDaoContractFactory) protected readonly contractFactory: QiDaoContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    const multicall = this.appToolkit.getMulticall(network);
+  getEscrowContract(address: string): QiDaoEscrowedQi {
+    return this.contractFactory.qiDaoEscrowedQi({ address, network: this.network });
+  }
 
-    const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
-    const address = '0x880decade22ad9c58a8a4202ef143c4f305100b3';
-    const qiToken = baseTokens.find(v => v.symbol === 'QI')!;
-    const tokens = [qiToken];
+  async getEscrowedTokenAddress({ contract }: GetTokenDefinitionsParams<QiDaoEscrowedQi>) {
+    return contract.Qi();
+  }
 
-    const qiTokencontract = this.qiDaoContractFactory.erc20({ address: qiToken.address, network });
-
-    const [decimals, balanceRaw] = await Promise.all([
-      multicall.wrap(qiTokencontract).decimals(),
-      multicall.wrap(qiTokencontract).balanceOf(address),
-    ]);
-
-    const balance = Number(balanceRaw) / 10 ** decimals;
-    const liquidity = balance * qiToken.price;
-
-    const contractPosition: ContractPosition = {
-      type: ContractType.POSITION,
-      address: address,
-      appId,
-      groupId,
-      network,
-      dataProps: {
-        liquidity,
-      },
-      displayProps: {
-        label: 'Escrowed QI',
-        secondaryLabel: buildDollarDisplayItem(qiToken.price),
-        images: [getTokenImg(qiToken.address, network)],
-        statsItems: [{ label: 'Liquidity', value: buildDollarDisplayItem(liquidity) }],
-      },
-      tokens,
-    };
-
-    return [contractPosition];
+  async getEscrowedTokenBalance({ address, contract }: GetTokenBalancesParams<QiDaoEscrowedQi>) {
+    return contract.userInfo(address).then(v => v.amount);
   }
 }