+++ src/apps/yearn/contracts/ethers/factories/index.ts
@@ -2,5 +2,8 @@
 /* tslint:disable */
 /* eslint-disable */
 export { YearnGovernance__factory } from './YearnGovernance__factory';
+export { YearnLpYCrv__factory } from './YearnLpYCrv__factory';
+export { YearnStakedYCrv__factory } from './YearnStakedYCrv__factory';
 export { YearnVault__factory } from './YearnVault__factory';
 export { YearnVaultV2__factory } from './YearnVaultV2__factory';
+export { YearnYCrv__factory } from './YearnYCrv__factory';